foodfinder/requirements.txt
asgiref==3.7.2
asttokens==2.4.1
decorator==5.1.1
Django==4.2.7
djangorestframework==3.14.0
exceptiongroup==1.2.0
executing==2.0.1
ipython==8.18.1
jedi==0.19.1
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.9.0
prompt-toolkit==3.0.41
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.0
typing_extensions==4.8.0
wcwidth==0.2.12
